Move ProxyConfiguration out of impl. package.
