+++ src/main/java/com/microsoft/azure/proton/transport/proxy/impl/StringUtils.java
@@ -1,7 +1,7 @@
 package com.microsoft.azure.proton.transport.proxy.impl;
 
-class StringUtils {
-    static boolean isNullOrEmpty(String string) {
+public class StringUtils {
+    public static boolean isNullOrEmpty(String string) {
         return string == null || string.isEmpty();
     }
 }
